xine-lib-1.2@11987

+++ src/video_out/video_out_xv.c
@@ -1262,8 +1262,8 @@ static vo_driver_t *open_plugin_2 (video_driver_class_t *class_gen, const void *
   if (xv_port != 0) {
     if (! xv_open_port(this, xv_port)) {
       xprintf(class->xine, XINE_VERBOSITY_NONE,
-	      _("%s: could not open Xv port %d - autodetecting\n"),
-	      LOG_MODULE, xv_port);
+	      _("%s: could not open Xv port %lu - autodetecting\n"),
+	      LOG_MODULE, (unsigned long)xv_port);
       xv_port = xv_autodetect_port(this, adaptors, adaptor_info, &adaptor_num, xv_port, prefer_type);
     } else
       adaptor_num = xv_find_adaptor_by_port (xv_port, adaptors, adaptor_info);
